+++ src/CrossValidation/Metrics/Informedness.php
@@ -88,17 +88,28 @@ class Informedness implements Metric
             }
         }
 
-        $score = 0.;
+        return array_sum(array_map(
+            [$this, 'compute'],
+            $truePositives,
+            $trueNegatives,
+            $falsePositives,
+            $falseNegatives
+        ))
+            / count($classes);
+    }
 
-        foreach ($truePositives as $class => $tp) {
-            $tn = $trueNegatives[$class];
-            $fp = $falsePositives[$class];
-            $fn = $falseNegatives[$class];
-
-            $score += $tp / (($tp + $fn) ?: self::EPSILON)
-                + $tn / (($tn + $fp) ?: self::EPSILON) - 1.;
-        }
-
-        return $score / count($classes);
+    /**
+     * Compute the class informedness score.
+     *
+     * @param int $tp
+     * @param int $tn
+     * @param int $fp
+     * @param int $fn
+     * @return float
+     */
+    public function compute(int $tp, int $tn, int $fp, int $fn) : float
+    {
+        return $tp / (($tp + $fn) ?: self::EPSILON)
+            + $tn / (($tn + $fp) ?: self::EPSILON) - 1.;
     }
 }